Move TracksListItem & TrackDecoration to TS
