+++ package.json
@@ -14,6 +14,7 @@
     "@babel/preset-react": "^7.0.0",
     "@sentry/browser": "^4.6.6",
     "@symfony/webpack-encore": "^0.27.0",
+    "@types/classnames": "^2.2.7",
     "@types/prop-types": "^15.7.1",
     "@types/react": "^16.8.18",
     "@types/react-dom": "^16.8.4",
@@ -37,7 +38,7 @@
     "react": "^16.8.3",
     "react-content-loader": "^3.4.1",
     "react-contextmenu": "^2.10.0",
-    "react-dnd": "^7.0.2",
+    "react-dnd": "^7.4.5",
     "react-dnd-html5-backend": "^7.0.2",
     "react-dom": "^16.6.3",
     "react-draggable": "^3.0.5",
@@ -60,5 +61,8 @@
     "start": "concurrently --kill-others \"yarn server\" \"yarn tsmp -w\"",
     "build": "yarn tsmp && yarn encore production",
     "tslint": "tslint -c tslint.json --project ."
+  },
+  "devDependencies": {
+    "musickit-typescript": "^1.2.4"
   }
 }
