+++ src/app/hooks/useMK.tsx
@@ -1,9 +1,9 @@
-import { useState, useEffect } from 'react';
+import { useEffect, useState } from 'react';
 
-export default function useMK(bindings = {}) {
-  const [events, setEvents] = useState({});
+export default function useMK(bindings: {[s: string]: string} = {}) {
+  const [events, setEvents] = useState<{[s: string]: string}>({});
 
-  function handleEventChange(key, e) {
+  function handleEventChange(key: string, e) {
     setEvents({
       ...events,
       [key]: e,
@@ -14,13 +14,15 @@ export default function useMK(bindings = {}) {
     const bindingFunctions = {};
 
     for (const [eventName, key] of Object.entries(bindings)) {
-      const handler = e => handleEventChange(key, e);
+      const handler = (e) => handleEventChange(key, e);
       bindingFunctions[eventName] = handler;
+      // @ts-ignore
       MusicKit.getInstance().addEventListener(eventName, handler);
     }
 
     return () => {
       for (const [eventName, func] of Object.entries(bindingFunctions)) {
+        // @ts-ignore
         MusicKit.getInstance().removeEventListener(eventName, func);
         delete bindingFunctions[eventName];
       }
